refactor: split internal/ftn/binkd.go along existing carve-outs
538 lines covering every binkd.conf operation, split to match the _write /
_sync / _parse carve-outs the package already uses, and to pair each half
with the test file that already targets it:

  binkd_write.go     206  fresh write, preserving rewrite, atomic file write
  binkd_sync.go      146  identity + per-link node sync (binkd_sync_test.go)
  binkd.go           123  config types, UpdateBinkdConf entry point
  binkd_settings.go   57  port/log-level sync (binkd_settings_test.go)
  binkd_parse.go      47  reading an existing conf

Pure move: no declaration body changed. Verified with a go/parser pairwise
diff of every top-level declaration (17 old, 17 new, 0 missing, 0 mismatched,
0 extra) and a comment multiset diff showing nothing removed. No test file
touched.
