+++ internal/ftn/binkd.go
@@ -4,9 +4,12 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"sort"
 	"strings"
 )
+
+// binkd.conf generation: the config types and the entry point that creates or
+// updates the file for a network. The writers, readers and per-link sync live
+// in the binkd_* files beside this one.
 
 // BinkdNode holds the parameters for a binkd node line.
 type BinkdNode struct {
@@ -117,422 +120,4 @@ func nodeExists(content, address string) bool {
 		}
 	}
 	return false
-}
-
-// writeFreshBinkdConf generates a complete binkd.conf from scratch.
-func writeFreshBinkdConf(out *strings.Builder, cfg BinkdConfig, outPath, logPath, secureIn, insecureIn, v3mailPath, boardName, sysop, location string) {
-	out.WriteString("# binkd.conf - Generated by FTN Setup Wizard\n\n")
-
-	// Domains, sorted so regeneration is deterministic despite the map.
-	out.WriteString("#\n# Domains:\n#\n")
-	names := make([]string, 0, len(cfg.Domains))
-	for name := range cfg.Domains {
-		names = append(names, name)
-	}
-	sort.Strings(names)
-	for _, name := range names {
-		fmt.Fprintf(out, "domain %s %s %d\n", name, outPath, cfg.Domains[name])
-	}
-
-	// Addresses.
-	out.WriteString("\n#\n# Your Addresses:\n#\n")
-	for _, addr := range cfg.Addresses {
-		fmt.Fprintf(out, "address %s\n", addr)
-	}
-
-	// General settings.
-	fmt.Fprintf(out, `
-#
-# General variables for your BBS:
-#
-sysname "%s"
-location "%s"
-sysop "%s"
-nodeinfo 9600,TCP,BINKP
-oblksize 4096
-timeout 60
-connect-timeout 0
-call-delay 300
-rescan-delay 60
-maxservers 4
-maxclients 1
-log %s
-loglevel 4
-conlog 4
-inbound %s
-inbound-nonsecure %s
-kill-dup-partial-files
-kill-old-partial-files 86400
-kill-old-bsy 43200
-try 2
-hold 1800
-iport 24554
-`, boardName, location, sysop, logPath, secureIn, insecureIn)
-
-	// Exec / prescan.
-	fmt.Fprintf(out, `
-#
-# Execute v3mail toss after receiving files:
-#
-prescan
-exec "%s toss" *.[mwtfsMWTFS][oehrauOEHRAU][0-9a-zA-Z] *.pkt
-
-#
-# BinkD additional options:
-#
-percents
-`, v3mailPath)
-}
-
-// rewriteBinkdConf filters an existing binkd.conf, replacing placeholder
-// lines with real values and injecting missing domain/address lines.
-func rewriteBinkdConf(out *strings.Builder, content string, cfg BinkdConfig, outPath, logPath, secureIn, insecureIn, v3mailPath, boardName, sysop, location string) {
-	// Collect domain names and addresses on lines that will be kept, so
-	// injection never duplicates them — binkd aborts on "duplicate domain".
-	// This matters when the real BBS root matches a placeholder path (e.g. a
-	// default /opt/vision3 install): the template's domain lines are then
-	// legitimate-looking and are kept rather than stripped.
-	kept := keptDirectives(content, cfg.BBSRoot)
-
-	writeDomains := func() {
-		for name, zone := range cfg.Domains {
-			if kept["domain "+name] {
-				continue
-			}
-			fmt.Fprintf(out, "domain %s %s %d\n", name, outPath, zone)
-		}
-	}
-	writeAddresses := func() {
-		for _, addr := range cfg.Addresses {
-			if kept["address "+addr] {
-				continue
-			}
-			fmt.Fprintf(out, "address %s\n", addr)
-		}
-	}
-
-	injectedDomains := false
-	injectedAddresses := false
-
-	// strings.Split rather than bufio.Scanner: a pathological over-64KB line
-	// would make the scanner stop early and silently drop the rest of the file.
-	for _, line := range confLines(content) {
-		trimmed := strings.TrimSpace(line)
-
-		// Strip placeholder lines entirely.
-		if isPlaceholderLine(line, cfg.BBSRoot) {
-			// If this is a domain placeholder, inject real domains once.
-			if strings.HasPrefix(trimmed, "domain ") && !injectedDomains {
-				writeDomains()
-				injectedDomains = true
-			}
-			// If this is an address placeholder, inject real addresses once.
-			if strings.HasPrefix(trimmed, "address ") && !injectedAddresses {
-				writeAddresses()
-				injectedAddresses = true
-			}
-			// Replace sysname/sysop/log/inbound/exec placeholders.
-			if strings.HasPrefix(trimmed, "sysname ") {
-				fmt.Fprintf(out, "sysname \"%s\"\n", boardName)
-				continue
-			}
-			if strings.HasPrefix(trimmed, "location ") {
-				fmt.Fprintf(out, "location \"%s\"\n", location)
-				continue
-			}
-			if strings.HasPrefix(trimmed, "sysop ") {
-				fmt.Fprintf(out, "sysop \"%s\"\n", sysop)
-				continue
-			}
-			if strings.HasPrefix(trimmed, "log ") {
-				fmt.Fprintf(out, "log %s\n", logPath)
-				continue
-			}
-			if strings.HasPrefix(trimmed, "inbound-nonsecure ") {
-				fmt.Fprintf(out, "inbound-nonsecure %s\n", insecureIn)
-				continue
-			}
-			if strings.HasPrefix(trimmed, "inbound ") {
-				fmt.Fprintf(out, "inbound %s\n", secureIn)
-				continue
-			}
-			if strings.HasPrefix(trimmed, "exec ") {
-				fmt.Fprintf(out, "exec \"%s toss\" *.[mwtfsMWTFS][oehrauOEHRAU][0-9a-zA-Z] *.pkt\n", v3mailPath)
-				continue
-			}
-			// Skip other placeholder lines (node hub.example.com, etc).
-			continue
-		}
-
-		// Keep non-placeholder lines as-is.
-		out.WriteString(line)
-		out.WriteByte('\n')
-	}
-
-	// If we never hit a domain/address placeholder, inject at the end —
-	// but only what isn't already present in kept lines.
-	if !injectedDomains && anyMissing(kept, "domain ", domainNames(cfg.Domains)) {
-		out.WriteString("\n#\n# Domains:\n#\n")
-		writeDomains()
-	}
-	if !injectedAddresses && anyMissing(kept, "address ", cfg.Addresses) {
-		out.WriteString("\n#\n# Your Addresses:\n#\n")
-		writeAddresses()
-	}
-}
-
-// keptDirectives returns a set of "domain <name>" and "address <addr>" keys
-// for non-placeholder lines in content — directives that a rewrite will keep.
-func keptDirectives(content, bbsRoot string) map[string]bool {
-	kept := make(map[string]bool)
-	for _, line := range confLines(content) {
-		if isPlaceholderLine(line, bbsRoot) {
-			continue
-		}
-		fields := strings.Fields(strings.TrimSpace(line))
-		if len(fields) >= 2 && (fields[0] == "domain" || fields[0] == "address") {
-			kept[fields[0]+" "+fields[1]] = true
-		}
-	}
-	return kept
-}
-
-// confLines splits conf content into lines without bufio.Scanner's 64KB
-// per-line limit; a trailing newline does not yield a spurious empty line.
-func confLines(content string) []string {
-	return strings.Split(strings.TrimSuffix(content, "\n"), "\n")
-}
-
-// domainNames returns the keys of a domain->zone map.
-func domainNames(domains map[string]int) []string {
-	names := make([]string, 0, len(domains))
-	for name := range domains {
-		names = append(names, name)
-	}
-	return names
-}
-
-// anyMissing reports whether any of names is absent from kept under prefix.
-func anyMissing(kept map[string]bool, prefix string, names []string) bool {
-	for _, n := range names {
-		if !kept[prefix+n] {
-			return true
-		}
-	}
-	return false
-}
-
-// BinkdIdentity holds BBS identity fields synced to binkd.conf.
-type BinkdIdentity struct {
-	BoardName string // sysname
-	SysopName string // sysop
-	Location  string // location
-}
-
-// BinkdLinkSync carries the per-link values synced into a binkd.conf node
-// line. HostPort is "hostname:port"; when empty only the password of an
-// existing line is synced and no new line is created (host unknown).
-type BinkdLinkSync struct {
-	SessionPwd string
-	HostPort   string
-}
-
-// SyncBinkdConf updates binkd.conf to reflect the current FTN links and BBS
-// identity fields (sysname, sysop, location). Node lines are upserted from
-// links: an existing line (matched by address) gets its hostname and password
-// refreshed, and a configured link with a hostname but no node line has one
-// appended — so a hub change in the TUI fully propagates. Only lines that
-// differ are rewritten; if nothing changed the file is not touched. Called
-// from saveAll so TUI edits are reflected automatically.
-//
-// links maps "address@network" (e.g. "21:1/100@fsxnet") to its sync values.
-func SyncBinkdConf(confPath string, identity BinkdIdentity, links map[string]BinkdLinkSync) error {
-	existing, err := os.ReadFile(confPath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return nil // No binkd.conf to sync.
-		}
-		return fmt.Errorf("reading binkd.conf: %w", err)
-	}
-
-	var out strings.Builder
-	changed := false
-	seenNodes := make(map[string]bool)
-
-	// confLines, not bufio.Scanner: an over-64KB line would stop a scanner
-	// early and the append path below would then persist a truncated file.
-	for _, line := range confLines(string(existing)) {
-		trimmed := strings.TrimSpace(line)
-
-		// Sync sysname.
-		if strings.HasPrefix(trimmed, "sysname ") && identity.BoardName != "" {
-			newLine := fmt.Sprintf("sysname \"%s\"", identity.BoardName)
-			if trimmed != newLine {
-				out.WriteString(newLine)
-				out.WriteByte('\n')
-				changed = true
-				continue
-			}
-		}
-
-		// Sync sysop.
-		if strings.HasPrefix(trimmed, "sysop ") && identity.SysopName != "" {
-			newLine := fmt.Sprintf("sysop \"%s\"", identity.SysopName)
-			if trimmed != newLine {
-				out.WriteString(newLine)
-				out.WriteByte('\n')
-				changed = true
-				continue
-			}
-		}
-
-		// Sync location.
-		if strings.HasPrefix(trimmed, "location ") && identity.Location != "" {
-			newLine := fmt.Sprintf("location \"%s\"", identity.Location)
-			if trimmed != newLine {
-				out.WriteString(newLine)
-				out.WriteByte('\n')
-				changed = true
-				continue
-			}
-		}
-
-		// Sync node hostname and session password.
-		if strings.HasPrefix(trimmed, "node ") {
-			fields := strings.Fields(trimmed)
-			if len(fields) >= 4 {
-				addr := fields[1] // e.g. "21:1/100@fsxnet"
-				if link, ok := links[addr]; ok {
-					seenNodes[addr] = true
-					newPwd := link.SessionPwd
-					if newPwd == "" {
-						newPwd = "-"
-					}
-					lineChanged := false
-					if link.HostPort != "" && fields[2] != link.HostPort {
-						fields[2] = link.HostPort
-						lineChanged = true
-					}
-					if fields[3] != newPwd {
-						fields[3] = newPwd
-						lineChanged = true
-					}
-					if lineChanged {
-						out.WriteString(strings.Join(fields, " "))
-						out.WriteByte('\n')
-						changed = true
-						continue
-					}
-				}
-			}
-		}
-
-		out.WriteString(line)
-		out.WriteByte('\n')
-	}
-
-	// Append node lines for configured links that have a hostname but no
-	// line yet (new link, or the link's address was changed in the TUI).
-	// Sorted so repeated syncs produce identical files.
-	var missing []string
-	for addr, link := range links {
-		if !seenNodes[addr] && link.HostPort != "" {
-			missing = append(missing, addr)
-		}
-	}
-	sort.Strings(missing)
-	for _, addr := range missing {
-		link := links[addr]
-		pwd := link.SessionPwd
-		if pwd == "" {
-			pwd = "-"
-		}
-		fmt.Fprintf(&out, "node %s %s %s\n", addr, link.HostPort, pwd)
-		changed = true
-	}
-
-	if !changed {
-		return nil
-	}
-	return writeFileAtomic(confPath, out.String(), 0600)
-}
-
-// SyncBinkdSettings updates the iport and loglevel lines in binkd.conf to
-// match the configured values. The file is only rewritten when a value
-// differs; a missing binkd.conf is a no-op (the FTN Setup Wizard creates it).
-// Non-positive port/logLevel values leave the corresponding line untouched.
-func SyncBinkdSettings(confPath string, port, logLevel int) error {
-	existing, err := os.ReadFile(confPath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return nil // No binkd.conf to sync.
-		}
-		return fmt.Errorf("reading binkd.conf: %w", err)
-	}
-
-	var out strings.Builder
-	changed := false
-
-	for _, line := range confLines(string(existing)) {
-		trimmed := strings.TrimSpace(line)
-
-		if strings.HasPrefix(trimmed, "iport ") && port > 0 {
-			newLine := fmt.Sprintf("iport %d", port)
-			if trimmed != newLine {
-				out.WriteString(newLine)
-				out.WriteByte('\n')
-				changed = true
-				continue
-			}
-		}
-		if strings.HasPrefix(trimmed, "loglevel ") && logLevel > 0 {
-			newLine := fmt.Sprintf("loglevel %d", logLevel)
-			if trimmed != newLine {
-				out.WriteString(newLine)
-				out.WriteByte('\n')
-				changed = true
-				continue
-			}
-		}
-
-		out.WriteString(line)
-		out.WriteByte('\n')
-	}
-
-	if !changed {
-		return nil
-	}
-	return writeFileAtomic(confPath, out.String(), 0600)
-}
-
-// writeFileAtomic creates the parent directory if needed and writes content via
-// a temp file + rename, so callers never see a partial/empty binkd.conf and
-// fresh installs (where data/ftn doesn't exist yet) don't fail.
-func writeFileAtomic(path, content string, perm os.FileMode) error {
-	if err := os.MkdirAll(filepath.Dir(path), 0755); err != nil {
-		return fmt.Errorf("creating dir for %s: %w", path, err)
-	}
-	tmp, err := os.CreateTemp(filepath.Dir(path), ".tmp.*")
-	if err != nil {
-		return fmt.Errorf("creating temp file: %w", err)
-	}
-	tmpName := tmp.Name()
-	if _, err := tmp.WriteString(content); err != nil {
-		_ = tmp.Close()        // cleanup on error path
-		_ = os.Remove(tmpName) // cleanup on error path
-		return fmt.Errorf("writing %s: %w", path, err)
-	}
-	if err := tmp.Close(); err != nil {
-		_ = os.Remove(tmpName) // cleanup on error path
-		return fmt.Errorf("closing %s: %w", path, err)
-	}
-	if err := os.Chmod(tmpName, perm); err != nil {
-		_ = os.Remove(tmpName) // cleanup on error path
-		return fmt.Errorf("chmod %s: %w", path, err)
-	}
-	if err := os.Rename(tmpName, path); err != nil {
-		_ = os.Remove(tmpName) // cleanup on error path
-		return fmt.Errorf("installing %s: %w", path, err)
-	}
-	return nil
 }
