+++ internal/ftn/binkd_write.go
@@ -1,0 +1,206 @@
+package ftn
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"sort"
+	"strings"
+)
+
+// Emitting binkd.conf content: a fresh file from scratch, a rewrite that
+// preserves directives the sysop added, and the atomic file write both use.
+
+// writeFreshBinkdConf generates a complete binkd.conf from scratch.
+func writeFreshBinkdConf(out *strings.Builder, cfg BinkdConfig, outPath, logPath, secureIn, insecureIn, v3mailPath, boardName, sysop, location string) {
+	out.WriteString("# binkd.conf - Generated by FTN Setup Wizard\n\n")
+
+	// Domains, sorted so regeneration is deterministic despite the map.
+	out.WriteString("#\n# Domains:\n#\n")
+	names := make([]string, 0, len(cfg.Domains))
+	for name := range cfg.Domains {
+		names = append(names, name)
+	}
+	sort.Strings(names)
+	for _, name := range names {
+		fmt.Fprintf(out, "domain %s %s %d\n", name, outPath, cfg.Domains[name])
+	}
+
+	// Addresses.
+	out.WriteString("\n#\n# Your Addresses:\n#\n")
+	for _, addr := range cfg.Addresses {
+		fmt.Fprintf(out, "address %s\n", addr)
+	}
+
+	// General settings.
+	fmt.Fprintf(out, `
+#
+# General variables for your BBS:
+#
+sysname "%s"
+location "%s"
+sysop "%s"
+nodeinfo 9600,TCP,BINKP
+oblksize 4096
+timeout 60
+connect-timeout 0
+call-delay 300
+rescan-delay 60
+maxservers 4
+maxclients 1
+log %s
+loglevel 4
+conlog 4
+inbound %s
+inbound-nonsecure %s
+kill-dup-partial-files
+kill-old-partial-files 86400
+kill-old-bsy 43200
+try 2
+hold 1800
+iport 24554
+`, boardName, location, sysop, logPath, secureIn, insecureIn)
+
+	// Exec / prescan.
+	fmt.Fprintf(out, `
+#
+# Execute v3mail toss after receiving files:
+#
+prescan
+exec "%s toss" *.[mwtfsMWTFS][oehrauOEHRAU][0-9a-zA-Z] *.pkt
+
+#
+# BinkD additional options:
+#
+percents
+`, v3mailPath)
+}
+
+// rewriteBinkdConf filters an existing binkd.conf, replacing placeholder
+// lines with real values and injecting missing domain/address lines.
+func rewriteBinkdConf(out *strings.Builder, content string, cfg BinkdConfig, outPath, logPath, secureIn, insecureIn, v3mailPath, boardName, sysop, location string) {
+	// Collect domain names and addresses on lines that will be kept, so
+	// injection never duplicates them — binkd aborts on "duplicate domain".
+	// This matters when the real BBS root matches a placeholder path (e.g. a
+	// default /opt/vision3 install): the template's domain lines are then
+	// legitimate-looking and are kept rather than stripped.
+	kept := keptDirectives(content, cfg.BBSRoot)
+
+	writeDomains := func() {
+		for name, zone := range cfg.Domains {
+			if kept["domain "+name] {
+				continue
+			}
+			fmt.Fprintf(out, "domain %s %s %d\n", name, outPath, zone)
+		}
+	}
+	writeAddresses := func() {
+		for _, addr := range cfg.Addresses {
+			if kept["address "+addr] {
+				continue
+			}
+			fmt.Fprintf(out, "address %s\n", addr)
+		}
+	}
+
+	injectedDomains := false
+	injectedAddresses := false
+
+	// strings.Split rather than bufio.Scanner: a pathological over-64KB line
+	// would make the scanner stop early and silently drop the rest of the file.
+	for _, line := range confLines(content) {
+		trimmed := strings.TrimSpace(line)
+
+		// Strip placeholder lines entirely.
+		if isPlaceholderLine(line, cfg.BBSRoot) {
+			// If this is a domain placeholder, inject real domains once.
+			if strings.HasPrefix(trimmed, "domain ") && !injectedDomains {
+				writeDomains()
+				injectedDomains = true
+			}
+			// If this is an address placeholder, inject real addresses once.
+			if strings.HasPrefix(trimmed, "address ") && !injectedAddresses {
+				writeAddresses()
+				injectedAddresses = true
+			}
+			// Replace sysname/sysop/log/inbound/exec placeholders.
+			if strings.HasPrefix(trimmed, "sysname ") {
+				fmt.Fprintf(out, "sysname \"%s\"\n", boardName)
+				continue
+			}
+			if strings.HasPrefix(trimmed, "location ") {
+				fmt.Fprintf(out, "location \"%s\"\n", location)
+				continue
+			}
+			if strings.HasPrefix(trimmed, "sysop ") {
+				fmt.Fprintf(out, "sysop \"%s\"\n", sysop)
+				continue
+			}
+			if strings.HasPrefix(trimmed, "log ") {
+				fmt.Fprintf(out, "log %s\n", logPath)
+				continue
+			}
+			if strings.HasPrefix(trimmed, "inbound-nonsecure ") {
+				fmt.Fprintf(out, "inbound-nonsecure %s\n", insecureIn)
+				continue
+			}
+			if strings.HasPrefix(trimmed, "inbound ") {
+				fmt.Fprintf(out, "inbound %s\n", secureIn)
+				continue
+			}
+			if strings.HasPrefix(trimmed, "exec ") {
+				fmt.Fprintf(out, "exec \"%s toss\" *.[mwtfsMWTFS][oehrauOEHRAU][0-9a-zA-Z] *.pkt\n", v3mailPath)
+				continue
+			}
+			// Skip other placeholder lines (node hub.example.com, etc).
+			continue
+		}
+
+		// Keep non-placeholder lines as-is.
+		out.WriteString(line)
+		out.WriteByte('\n')
+	}
+
+	// If we never hit a domain/address placeholder, inject at the end —
+	// but only what isn't already present in kept lines.
+	if !injectedDomains && anyMissing(kept, "domain ", domainNames(cfg.Domains)) {
+		out.WriteString("\n#\n# Domains:\n#\n")
+		writeDomains()
+	}
+	if !injectedAddresses && anyMissing(kept, "address ", cfg.Addresses) {
+		out.WriteString("\n#\n# Your Addresses:\n#\n")
+		writeAddresses()
+	}
+}
+
+// writeFileAtomic creates the parent directory if needed and writes content via
+// a temp file + rename, so callers never see a partial/empty binkd.conf and
+// fresh installs (where data/ftn doesn't exist yet) don't fail.
+func writeFileAtomic(path, content string, perm os.FileMode) error {
+	if err := os.MkdirAll(filepath.Dir(path), 0755); err != nil {
+		return fmt.Errorf("creating dir for %s: %w", path, err)
+	}
+	tmp, err := os.CreateTemp(filepath.Dir(path), ".tmp.*")
+	if err != nil {
+		return fmt.Errorf("creating temp file: %w", err)
+	}
+	tmpName := tmp.Name()
+	if _, err := tmp.WriteString(content); err != nil {
+		_ = tmp.Close()        // cleanup on error path
+		_ = os.Remove(tmpName) // cleanup on error path
+		return fmt.Errorf("writing %s: %w", path, err)
+	}
+	if err := tmp.Close(); err != nil {
+		_ = os.Remove(tmpName) // cleanup on error path
+		return fmt.Errorf("closing %s: %w", path, err)
+	}
+	if err := os.Chmod(tmpName, perm); err != nil {
+		_ = os.Remove(tmpName) // cleanup on error path
+		return fmt.Errorf("chmod %s: %w", path, err)
+	}
+	if err := os.Rename(tmpName, path); err != nil {
+		_ = os.Remove(tmpName) // cleanup on error path
+		return fmt.Errorf("installing %s: %w", path, err)
+	}
+	return nil
+}
